+++ Bezier.cpp
@@ -1,6 +1,6 @@
 #include "Bezier.h"
 
-const double Bezier::CUBIC_FITTING_ERROR  = 0.000099;
+const FLOAT Bezier::CUBIC_FITTING_ERROR  = 0.000099;
 
 Bezier::Bezier(void)
 {
@@ -18,16 +18,16 @@ void Bezier::cubicBezierLeastsquare(const CVector3f *q,const int n,CVector3f &p1
     assert(n > 3);
     CVector3f p0 = q[0],p3 = q[n-1];
 
-    double A1 = 0,A2 = 0, A12 = 0;
+    FLOAT A1 = 0,A2 = 0, A12 = 0;
     CVector3f C1(0,0,0),C2(0,0,0);
 
     for(int i = 0;i < n;i++)
     {
-        double t    = i*1.0/(n-1);
-        double t2   = t*t , t3 = t2*t , t4   = t3*t;
+        FLOAT t    = i*1.0/(n-1);
+        FLOAT t2   = t*t , t3 = t2*t , t4   = t3*t;
 
-        double _t   = 1-t ;
-        double _t2  = _t*_t , _t3  = _t2*_t, _t4  = _t3*_t;
+        FLOAT _t   = 1-t ;
+        FLOAT _t2  = _t*_t , _t3  = _t2*_t, _t4  = _t3*_t;
 
         A1 += t2 * _t4;
         A2 += t4 * _t2;
@@ -42,31 +42,31 @@ void Bezier::cubicBezierLeastsquare(const CVector3f *q,const int n,CVector3f &p1
     //P1 = (A2C1-A12C2)/(A1A2-A12A12)
     //P2 = (A1C2-A12C1)/(A1A2-A12A12)
     A1 *= 9; A2 *= 9; A12 *= 9;
-    double base = A1*A2 - A12*A12;
+    FLOAT base = A1*A2 - A12*A12;
     assert(base != 0);
     p1 = (A2*C1 - A12*C2)/base;
     p2 = (A1*C2 - A12*C1)/base;
 }
 
-void Bezier::cubicBezierLeastsquare(const double *q,const int n,const int dim,double *p1,double *p2)
+void Bezier::cubicBezierLeastsquare(const FLOAT *q,const int n,const int dim,FLOAT *p1,FLOAT *p2)
 {
     assert(n > 3);
-    const double *p0 = &q[0*dim],*p3 = &q[(n-1)*dim];
+    const FLOAT *p0 = &q[0*dim],*p3 = &q[(n-1)*dim];
 
-    double A1 = 0,A2 = 0, A12 = 0;
-	double *C1 = new double[dim]; memset(C1,0,sizeof(double)*dim);
-	double *C2 = new double[dim]; memset(C2,0,sizeof(double)*dim);
+    FLOAT A1 = 0,A2 = 0, A12 = 0;
+	FLOAT *C1 = new FLOAT[dim]; memset(C1,0,sizeof(FLOAT)*dim);
+	FLOAT *C2 = new FLOAT[dim]; memset(C2,0,sizeof(FLOAT)*dim);
 
 	#ifndef __STDC_IEC_559__
-	assert(1 == 0);/*Maybe there are some problems with memset, which is used to initialize double value */
+	assert(1 == 0);/*Maybe there are some problems with memset, which is used to initialize FLOAT value */
 	#endif
 
     for(int i = 0;i < n;i++)
     {
-        double t = i*1.0/(n-1);
-        double t2   = t*t , t3 = t2*t , t4   = t3*t;
-        double _t   = 1-t;
-		double _t2  = _t*_t , _t3  = _t2*_t, _t4  = _t3*_t;
+        FLOAT t = i*1.0/(n-1);
+        FLOAT t2   = t*t , t3 = t2*t , t4   = t3*t;
+        FLOAT _t   = 1-t;
+		FLOAT _t2  = _t*_t , _t3  = _t2*_t, _t4  = _t3*_t;
 
         A1 += t2 * _t4;
         A2 += t4 * _t2;
@@ -74,7 +74,7 @@ void Bezier::cubicBezierLeastsquare(const double *q,const int n,const int dim,do
 
 		for(int j = 0;j < dim;j++)
 		{
-			double tmp = (q[i*dim+j] - _t3*p0[j] - t3*p3[j]);
+			FLOAT tmp = (q[i*dim+j] - _t3*p0[j] - t3*p3[j]);
 			C1[j] += ( 3*t*_t2*tmp );
 			C2[j] += ( 3*t2*_t*tmp );
 		}
@@ -83,7 +83,7 @@ void Bezier::cubicBezierLeastsquare(const double *q,const int n,const int dim,do
     //P1 = (A2C1-A12C2)/(A1A2-A12A12)
     //P2 = (A1C2-A12C1)/(A1A2-A12A12)
     A1 *= 9; A2 *= 9; A12 *= 9;
-    double base = A1*A2 - A12*A12;
+    FLOAT base = A1*A2 - A12*A12;
     assert(base != 0);
 	for(int j = 0;j < dim;j++)
 	{
@@ -110,16 +110,16 @@ int Bezier::cubicBezierFitting(const CVector3f *q,const int n,const CVector3f &P
     a.x = P3.x - P0.x - c.x - b.x;
     a.y = P3.y - P0.y - c.y - b.y;
     a.z = P3.z - P0.z - c.z - b.z;
-    double max_error = 0; int max_id = -1;
+    FLOAT max_error = 0; int max_id = -1;
 
     for(int i = 1;i < n-1;i++)//zero-based
     {
-        double t = i*1.0/(n-1);
+        FLOAT t = i*1.0/(n-1);
         p.x = a.x * t * t * t + b.x * t * t + c.x * t + P0.x;
         p.y = a.y * t * t * t + b.y * t * t + c.y * t + P0.y;
         p.z = a.z * t * t * t + b.z * t * t + c.z * t + P0.z;
 
-        double dist = manhattanDist(q[i],p)/3;//sum / 3
+        FLOAT dist = manhattanDist(q[i],p)/3;//sum / 3
 
         if(H3DMath::complf(dist - max_error) > 0)
         {
@@ -133,14 +133,14 @@ int Bezier::cubicBezierFitting(const CVector3f *q,const int n,const CVector3f &P
     }
     return -1;
 }
-int Bezier::cubicBezierFitting(const double *q,const int n,const int dim,const double *P1,const double *P2)
+int Bezier::cubicBezierFitting(const FLOAT *q,const int n,const int dim,const FLOAT *P1,const FLOAT *P2)
 {
-	const double *P0 = &q[0*dim],*P3 = &q[(n-1)*dim];
+	const FLOAT *P0 = &q[0*dim],*P3 = &q[(n-1)*dim];
     //p(t) = P0 * (1-t)^3 + P1 * 3*t(1-t)^2 + P2 * 3*t^2*(1-t) + P3 * t^3
-    double *a = new double[dim];
-	double *b = new double[dim];
-	double *c = new double[dim];
-	double p;
+    FLOAT *a = new FLOAT[dim];
+	FLOAT *b = new FLOAT[dim];
+	FLOAT *c = new FLOAT[dim];
+	FLOAT p;
 	for(int j = 0;j < dim;j++)
 	{
 		c[j] = 3 * (P1[j] - P0[j]);
@@ -148,11 +148,11 @@ int Bezier::cubicBezierFitting(const double *q,const int n,const int dim,const d
 		a[j] = P3[j] - P0[j] - c[j] - b[j];
 	}
 
-    double max_error = 0; int max_id = -1;
+    FLOAT max_error = 0; int max_id = -1;
 
     for(int i = 1;i < n-1;i++)//zero-based
     {
-        double t = i*1.0/(n-1);
+        FLOAT t = i*1.0/(n-1);
 		int index = i*dim;
 		for(int j = 0;j < dim;j++)
 		{
@@ -191,19 +191,19 @@ void Bezier::cubicBezierInterp(CVector3f *q,const int n,const CVector3f &P1,cons
 
     for(int i = 1;i < n-1;i++)
     {
-        double t = i*1.0/(n-1);
+        FLOAT t = i*1.0/(n-1);
         q[i].x = a.x * t * t * t + b.x * t * t + c.x * t + P0.x;
         q[i].y = a.y * t * t * t + b.y * t * t + c.y * t + P0.y;
         q[i].z = a.z * t * t * t + b.z * t * t + c.z * t + P0.z;
     }
 }
-void Bezier::cubicBezierInterp(double *q,const int n,const int dim,const double *P1,const double *P2)
+void Bezier::cubicBezierInterp(FLOAT *q,const int n,const int dim,const FLOAT *P1,const FLOAT *P2)
 {
-	const double *P0 = &q[0*dim],*P3 = &q[(n-1)*dim];
+	const FLOAT *P0 = &q[0*dim],*P3 = &q[(n-1)*dim];
     //p(t) = P0 * (1-t)^3 + P1 * 3*t(1-t)^2 + P2 * 3*t^2*(1-t) + P3 * t^3
-    double *a = new double[dim];
-	double *b = new double[dim];
-	double *c = new double[dim];
+    FLOAT *a = new FLOAT[dim];
+	FLOAT *b = new FLOAT[dim];
+	FLOAT *c = new FLOAT[dim];
 
 	for(int j = 0;j < dim;j++)
 	{
@@ -214,7 +214,7 @@ void Bezier::cubicBezierInterp(double *q,const int n,const int dim,const double 
 
     for(int i = 1;i < n-1;i++)//zero-based
     {
-        double t = i*1.0/(n-1);
+        FLOAT t = i*1.0/(n-1);
 		int index = i*dim;
 		for(int j = 0;j < dim;j++)
 		{
